chore: refactor with modules for each subsystem
